Allow numpy log(0) warnings; State reproducibility caveats.
Slight differences in results across machines and systems are expected due to float-point calculations. Spherical KMeans can fail to cluster a specific problem even with true k due to relying on a set of assumptions; In that case, one cluster is returned, and that is taken into account and saved to disk as k-vals csv too.